+++ .env.sample
@@ -9,3 +9,4 @@ TRELLO_KEY=
 TRELLO_OAUTH_SECRET=
 
 # note: .env is a shell file so there can't be spaces around =
+# note: rename .env.sample to .env
